requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2

streamlit==1.44.0
requests==2.32.3

numpy==1.26.4
pillow==10.4.0
pycryptodome==3.20.0
scikit-image==0.24.0

pytest==8.3.3